test: AppUserService와 TxUserServiceTest로 테스트 메서드 분리
