+++ app/src/test/java/com/techcourse/service/AppUserServiceTest.java
@@ -1,0 +1,44 @@
+package com.techcourse.service;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
+
+import nextstep.jdbc.JdbcTemplate;
+
+class AppUserServiceTest {
+
+    private JdbcTemplate jdbcTemplate;
+    private UserDao userDao;
+
+    @BeforeEach
+    void setUp() {
+        this.jdbcTemplate = new JdbcTemplate(DataSourceConfig.getInstance());
+        this.userDao = new UserDao(jdbcTemplate);
+
+        DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
+        User user = new User("gugu", "password", "hkkang@woowahan.com");
+        userDao.insert(user);
+    }
+
+    @Test
+    void testChangePassword() {
+        UserHistoryDao userHistoryDao = new UserHistoryDao(jdbcTemplate);
+        UserService userService = new AppUserService(userDao, userHistoryDao);
+
+        String newPassword = "qqqqq";
+        String createBy = "gugu";
+        userService.changePassword(1L, newPassword, createBy);
+
+        User actual = userService.findById(1L);
+
+        assertThat(actual.getPassword()).isEqualTo(newPassword);
+    }
+}
